requirements.txt
requests==2.31.0
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
websocket-client==1.8.0
celery==5.3.6
redis==5.0.1
uvicorn==0.29.0
pydantic==2.7.1
fastapi==0.111.0
motor==3.4.0
boto3==1.34.127
cryptography==42.0.8
